convert useDarkMode to ts
